moved list of records to own home.jsx component
